Extract shared `get_first_post_body` helper; move Bedrock cache test to its own file
Wave-2 review follow-ups:
- Pull the "first POST body parsed as JSON" walk out of `get_tool_choice_from_cassette` and `get_bedrock_tool_config_from_cassette` into a single `get_first_post_body` helper in `tests/cassette_utils.py`. Handles both raw bytes/str and pre-parsed dict bodies (some VCR serializers, e.g. the project's huggingface JSON-body serializer, deserialize ahead of playback).
- Move `test_bedrock_single_tool_choice_preserves_cache` from the legacy per-provider `tests/models/test_bedrock.py` into a new feature-centric `tests/models/test_bedrock_cache.py`. Cassettes follow via `git mv` to `tests/models/cassettes/test_bedrock_cache/`.
