+++ tests/models/cassettes/test_bedrock_cache/test_bedrock_single_tool_choice_preserves_cache[nova].yaml
